requirements.txt
asgiref==3.2.10
astroid==2.3.3
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
deepdiff==5.0.2
Django==3.0.6
django-cors-headers==3.3.0
django-debug-toolbar==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.5
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
iniconfig==1.0.1
isort==4.3.21
jsondiff==1.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
ordered-set==4.0.2
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.1.0
pytest-django==3.10.0
pytz==2020.1
rfc3986==1.4.0
service-identity==18.1.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
uvicorn==0.12.1
uvloop==0.14.0
wrapt==1.11.2
zope.interface==5.1.1
